+++ R/mspline.R
@@ -19,7 +19,7 @@
 #'   mean survival will be conditioned on survival up to this time
 #'
 #' @details Survival models with a flexible M-spline on the baseline hazard are
-#'   described by \insertCite{Brilleman2020;textual}{multinma}.
+#'   described by \insertCite{Phillippo_mspline;textual}{multinma}.
 #'   Piecewise-exponential baseline hazards are a special case where the degree
 #'   of the M-spline polynomial is 0.
 #'